Refactors Delauanay class into several detail files separating concerns
